+++ presentation/src/main/res/layout/activity_main.xml
@@ -19,6 +19,7 @@
             android:id="@+id/appbar_layout"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
+            android:visibility="gone"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent">
@@ -62,8 +63,6 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             />
-
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </layout>
