+++ presentation/src/main/res/layout/fragment_search_main.xml
@@ -29,6 +29,14 @@
 <!--        app:defaultQueryHint="@string/search_main_hint"-->
 
 
+        <include
+            android:id="@+id/toolbar_search_main"
+            layout="@layout/toolbar_search_main"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"/>
 
 
         <ScrollView
@@ -37,7 +45,7 @@
             android:layout_height="0dp"
             android:fillViewport="true"
 
-            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/toolbar_search_main"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
